Work without requiring ActiveRecord.
This allows using it with other ORMs like Sequel, and in non-Rails
projects. Transactions, however, only make sense with AR, so an error is
raised when using them without it.

This moves AR to development dependencies, adn adds sqlite there, only
for testing purposes.

Additionally, a custom RSpec matcher is invented to check broadcast
responses, it may be useful to simplify other examples.